Enable printing the build info from the rocshmem library itself
