+++ projects/rocshmem/CMakeLists.txt
@@ -220,6 +220,17 @@ if (NOT BUILD_TESTS_ONLY)
 
   configure_file(cmake/rocshmem_config.h.in include/rocshmem/rocshmem_config.h)
 
+  # Embed rocshmem_config.h as a C++17 raw string literal so build_info.cpp
+  # (and the rocshmem_info tool) can display it without reading the file at runtime.
+  file(READ "${CMAKE_BINARY_DIR}/include/rocshmem/rocshmem_config.h" _config_content)
+  file(WRITE "${CMAKE_BINARY_DIR}/include/rocshmem/rocshmem_config_embedded.hpp"
+    "// Auto-generated from rocshmem_config.h — do not edit.\n"
+    "#pragma once\n"
+    "static const char rocshmem_config_h_content[] = R\"ROCSHMEM_CFG_END(\n"
+    "${_config_content}"
+    ")ROCSHMEM_CFG_END\";\n"
+  )
+
   #############################################################################
   # LINKING AND INCLUDE DIRECTORIES
   #############################################################################
