+++ projects/rocshmem/src/build_info.hpp
@@ -1,0 +1,44 @@
+/******************************************************************************
+ * Copyright (c) Advanced Micro Devices, Inc. All rights reserved.
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to
+ * deal in the Software without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+ * sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+ * IN THE SOFTWARE.
+ *****************************************************************************/
+
+#ifndef LIBRARY_SRC_BUILD_INFO_HPP_
+#define LIBRARY_SRC_BUILD_INFO_HPP_
+
+#include <iostream>
+
+namespace rocshmem {
+
+/**
+ * Print a formatted block showing the rocSHMEM build configuration:
+ * version, git hash, install prefix, GPU arch, ROCm/MPI versions, and the
+ * compile-time flags from rocshmem_config.h.
+ *
+ * Called by the rocshmem_info tool and by the library itself when
+ * ROCSHMEM_DEBUG_LEVEL >= INFO.
+ */
+void print_build_info(std::ostream& os = std::cout);
+
+}  // namespace rocshmem
+
+#endif  // LIBRARY_SRC_BUILD_INFO_HPP_
